Fix 19 additional bugs: NIF parser safety, BSA underflow, collision division-by-zero, temp file leaks
Critical fixes:
- nif_parser.cpp: readString() now accepts strings up to 1MB (was 256 bytes)
- nif_parser.cpp: readHeader() reads magic until newline instead of fixed 256 bytes
- nif_parser.cpp: Add MAX_PARTITIONS/MAX_KEYS/MAX_TEXT_KEYS limits to prevent OOM
- bsa_reader.cpp: Guard against entry.size < 4 uint32_t underflow in extractFileDecompressed
- collision_world.cpp: Guard division-by-zero when b_dot/a_dot is 0 in npCapsuleCapsule
- native_activity.cpp: Set window_changed=true in onDestroy to wake render thread
- native_activity.cpp: Use actual frame time instead of fixed 0.0167f delta

High fixes:
- asset_manager.cpp: Delete temp NIF/DDS files after parsing
- asset_manager.cpp: Update lastAccessTime in update() for proper LRU eviction
- asset_manager.cpp: Normalize path separators in loadNifMesh fallback
- ui_manager.cpp: Fix self-assignment (textRenderer = textRenderer -> this->textRenderer)
- mesh.cpp: Use uintptr_t for safe 64-bit pointer arithmetic in buildPartitionRenderInfo
- title_screen.cpp: Remove redundant glClear in renderOblivionLogo

+++ app/src/main/cpp/geometry/mesh.cpp
@@ -149,7 +149,9 @@ void SkinnedMesh::buildPartitionRenderInfo() {
     uint32_t offset = 0;
     for (const auto& p : partitions) {
         partitionCounts.push_back(static_cast<GLsizei>(p.numTriangles * 3));
-        partitionOffsets.push_back(reinterpret_cast<const void*>(offset * sizeof(uint16_t)));
+        // FIX: Use uintptr_t for safe 64-bit pointer arithmetic
+        partitionOffsets.push_back(reinterpret_cast<const void*>(
+            static_cast<uintptr_t>(offset) * sizeof(uint16_t)));
         offset += p.numTriangles * 3;
     }
 }
